# docs/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: University Faculty Work Portal API
  description: |
    API для системи планування, звітування та оцінювання роботи викладачів університету.
    
    **Типи користувачів:**
    - GUEST - незареєстрований користувач (публічний доступ)
    - TEACHER - викладач
    - DIRECTOR - завідувач кафедри/декан (має права викладача + додаткові)
    - ADMIN - адміністратор системи
    
    **Автентифікація:**
    - JWT tokens (Access: 30 хв, Refresh: 30 днів)
    - Hybrid strategy з Redis кешуванням
  version: 1.0.0
  contact:
    name: API Support
    email: api@stu.cn.ua
servers:
  - url: https://npp.stu.cn.ua/api/v1
    description: Production server
  - url: http://localhost:8080/v1
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PublicActivityResponse:
      type: object
      description: "Публічна версія роботи - без коефіцієнтів та службових полів"
      properties:
        id:
          type: integer
        description:
          type: string
        year:
          type: integer
        semester:
          type: integer
          description: "0 - весь рік, 1 - перший семестр, 2 - другий семестр"
        worksType:
          type: string
          enum: [SCIENCE, METHOD, ORG]
        createdAt:
          type: string
          format: date-time
        users:
          type: array
          description: "Автори роботи (тільки ПІБ, без деталей)"
          items:
            type: object
            properties:
              id:
                type: integer
              firstName:
                type: string
              lastName:
                type: string
              patronymic:
                type: string
...
    FacultyResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        facLogo:
          type: string
        facHead:
          type: string
        facPage:
          type: string
        active:
          type: boolean
        institute:
          $ref: '#/components/schemas/InstituteResponse'
    CreateFacultyRequest:
      type: object
      required:
        - name
        - instituteId
        - facHead
      properties:
        name:
          type: string
        instituteId:
          type: integer
        facLogo:
          type: string
        facHead:
          type: string
        facPage:
          type: string
    UpdateFacultyRequest:
      type: object
      properties:
        name:
          type: string
        instituteId:
          type: integer
        facLogo:
          type: string
        facHead:
          type: string
        facPage:
          type: string
    InstituteResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        instLogo:
          type: string
        instHead:
          type: string
        instPage:
          type: string
        active:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
paths:
...
  /faculties:
    get:
      tags:
        - Faculties
      summary: Список факультетів
      operationId: getFaculties
      parameters:
        - name: instituteId
          in: query
          schema:
            type: integer
        - name: active
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Список факультетів
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FacultyResponse'

    post:
      tags:
        - Faculties
      summary: Створення факультету
      description: |
        Створення нового факультету.
        **Доступ:** тільки ADMIN
      operationId: createFaculty
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFacultyRequest'
      responses:
        '201':
          description: Факультет створено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacultyResponse'
        '401':
          description: Не авторизований або токен недійсний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостатньо прав (тільки ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /faculties/{id}:
    get:
      tags:
        - Faculties
      summary: Деталі факультету
      operationId: getFacultyById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Деталі факультету
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacultyResponse'
        '404':
          description: Факультет не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Faculties
      summary: Оновлення факультету
      description: |
        **Доступ:** тільки ADMIN
      operationId: updateFaculty
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFacultyRequest'
      responses:
        '200':
          description: Факультет оновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacultyResponse'
        '401':
          description: Не авторизований або токен недійсний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостатньо прав (тільки ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Факультет не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Faculties
      summary: Видалення факультету
      description: |
        **Доступ:** тільки ADMIN
      operationId: deleteFaculty
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Факультет видалено
        '401':
          description: Не авторизований або токен недійсний
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Недостатньо прав (тільки ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Факультет не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== INSTITUTES ====================
...
  /public/teacher/{teacherId}/activities:
    get:
      tags:
        - Public
      summary: Публічні роботи викладача
      description: |
        Список робіт викладача (доступний гостям).
        **Без коефіцієнтів та службової інформації.**
      operationId: getPublicTeacherActivities
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: integer
        - name: worksType
          in: query
          schema:
            type: string
            enum: [SCIENCE, METHOD, ORG]
        - name: year
          in: query
          schema:
            type: integer
        - name: semester
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Список робіт (публічна версія)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicActivityResponse'
        '404':
          description: Викладача не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...